Data.Hashable.Color

Description

Hashes are an easy way to distinguish between two objects since it is not impossible that two random objects have the same hash, but it is very unlikely. But for humans, a hash is often still not very convenient. Colors are an easier way to distinguish. For example one can make a list of items, and give each of the items a different color based on the hash.

This module provides a function rgbHash that can convert any Hashable object to a Colour.

Determine color

rgbHash

Arguments

:: (Hashable a, Floating b, Ord b)

=> a
The Hashable object to convert to a color.

-> Colour b
The corresponding color to use.

Convert a given Hashable object to a Colour by determining the hash, and using the last 24 bits as source for the red, green, and blue channels of the Colour to construct.

Show items with color

rgbHashBgShow

Arguments

:: (Hashable a, Show a)

=> a
The Hashable object to print in a colorized way.

-> String
A String with ANSI terminal codes to colorize the object

Show the given Hashable object with ANSI terminal codes such that the background is colorized by the hash of that object. This will only work if the terminal supports 24-bit colors. The foreground is either black or white depending on whether the color is more light or dark.

rgbHashFgShow

Arguments

:: (Hashable a, Show a)

=> a
The Hashable object to print in a colorized way.

-> String
A String with ANSI terminal codes to colorize the object

Show the given Hashable object with ANSI terminal codes such that the foreground is colorized by the hash of that object. This will only work if the terminal supports 24-bit colors. The background is either black or white depending on whether the color is more light or dark.